+++ meerschaum/core/Pipe/_fetch.py
@@ -8,18 +8,20 @@ Functions for fetching new data into the Pipe
 
 from __future__ import annotations
 from datetime import timedelta, datetime
-from meerschaum.utils.typing import Optional, Any, Union
+import meerschaum as mrsm
+from meerschaum.utils.typing import Optional, Any, Union, SuccessTuple, Iterator
 from meerschaum.config import get_config
+from meerschaum.utils.warnings import warn
 
 def fetch(
         self,
         begin: Optional[datetime.datetime, str] = '',
         end: Optional[datetime.datetime] = None,
+        check_existing: bool = True,
         sync_chunks: bool = False,
-        deactivate_plugin_venv: bool = True,
         debug: bool = False,
         **kw: Any
-    ) -> 'pd.DataFrame or None':
+    ) -> Union['pd.DataFrame', Iterator['pd.DataFrame'], None]:
     """
     Fetch a Pipe's latest data from its connector.
 
@@ -30,6 +32,9 @@ def fetch(
 
     end: Optional[datetime.datetime], default None:
         If provided, only fetch data older than or equal to `end`.
+
+    check_existing: bool, default True
+        If `False`, do not apply the backtrack interval.
 
     sync_chunks: bool, default False
         If `True` and the pipe's connector is of type `'sql'`, begin syncing chunks while fetching
@@ -44,27 +49,15 @@ def fetch(
 
     """
     if 'fetch' not in dir(self.connector):
-        from meerschaum.utils.warnings import warn
         warn(f"No `fetch()` function defined for connector '{self.connector}'")
         return None
 
-    from meerschaum.connectors import custom_types
+    import contextlib
+    from meerschaum.connectors import custom_types, get_connector_plugin
     from meerschaum.utils.debug import dprint, _checkpoint
-    if (
-        self.connector.type == 'plugin'
-        or
-        self.connector.type in custom_types
-    ):
-        from meerschaum.plugins import Plugin
-        from meerschaum.utils.packages import activate_venv, deactivate_venv
-        plugin_name = (
-            self.connector.label if self.connector.type == 'plugin'
-            else self.connector.__module__.replace('plugins.', '').split('.')[0]
-        )
-        connector_plugin = Plugin(plugin_name)
-        connector_plugin.activate_venv(debug=debug)
-    
+
     _chunk_hook = kw.pop('chunk_hook', None)
+    kw['workers'] = self.get_num_workers(kw.get('workers', None))
     if sync_chunks and _chunk_hook is None:
 
         def _chunk_hook(chunk, **_kw) -> SuccessTuple:
@@ -79,25 +72,36 @@ def fetch(
                 chunk_message = '\n' + chunk_label + '\n' + chunk_message
             return chunk_success, chunk_message
 
-    workers = kw.get('workers', None)
-    if workers is None and not getattr(self.instance_connector, 'IS_THREAD_SAFE', False):
-        workers = 1
-    kw['workers'] = workers
 
-    df = self.connector.fetch(
-        self,
-        begin = _determine_begin(self, begin, debug=debug),
-        end = end,
-        chunk_hook = _chunk_hook,
-        debug = debug,
-        **kw
-    )
+    with mrsm.Venv(get_connector_plugin(self.connector)):
+        df = self.connector.fetch(
+            self,
+            begin = _determine_begin(
+                self,
+                begin,
+                check_existing = check_existing,
+                debug = debug,
+            ),
+            end = end,
+            chunk_hook = _chunk_hook,
+            debug = debug,
+            **kw
+        )
     return df
 
 
-def get_backtrack_interval(self, debug: bool = False) -> Union[timedelta, int]:
+def get_backtrack_interval(
+        self,
+        check_existing: bool = True,
+        debug: bool = False,
+    ) -> Union[timedelta, int]:
     """
     Get the chunk interval to use for this pipe.
+
+    Parameters
+    ----------
+    check_existing: bool, default True
+        If `False`, return a backtrack_interval of 0 minutes.
 
     Returns
     -------
@@ -109,7 +113,7 @@ def get_backtrack_interval(self, debug: bool = False) -> Union[timedelta, int]:
         configured_backtrack_minutes
         if configured_backtrack_minutes is not None
         else default_backtrack_minutes
-    )
+    ) if check_existing else 0
 
     backtrack_interval = timedelta(minutes=backtrack_minutes)
     dt_col = self.columns.get('datetime', None)
@@ -124,19 +128,33 @@ def get_backtrack_interval(self, debug: bool = False) -> Union[timedelta, int]:
 
 
 def _determine_begin(
-        pipe: meerschaum.Pipe,
+        pipe: mrsm.Pipe,
         begin: Union[datetime, int, str] = '',
+        check_existing: bool = True,
         debug: bool = False,
     ) -> Union[datetime, int, None]:
     """
     Apply the backtrack interval if `--begin` is not provided.
+
+    Parameters
+    ----------
+    begin: Union[datetime, int, str], default ''
+        The provided begin timestamp.
+
+    check_existing: bool, default True
+        If `False`, do not apply the backtrack interval.
+
+    Returns
+    -------
+    A datetime (or int) value from which to fetch.
+    Returns `None` if no begin may be determined.
     """
     if begin != '':
         return begin
     sync_time = pipe.get_sync_time(debug=debug)
     if sync_time is None:
         return sync_time
-    backtrack_interval = pipe.get_backtrack_interval(debug=debug)
+    backtrack_interval = pipe.get_backtrack_interval(check_existing=check_existing, debug=debug)
     if isinstance(sync_time, datetime) and isinstance(backtrack_interval, int):
         backtrack_interval = timedelta(minutes=backtrack_interval)
     try: